requirements.txt
junit_xml == 1.9
grpcio == 1.46.1
grpcio-tools == 1.46.1
protobuf == 3.20.1
